+++ src/Dap.Build/RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 * Generate .sha512 file when inject
 * Save original version before inject
 * Create Recover target to remove the injected version
+* Create sub tragets per project
 
 ## 0.6.2
 * Create targets for each projects